#146 [FIX] client 페이지 분리
